Phase two of Architecture rebuild.
