--- pages/cameraTest.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Camera Testing</title>
-
-    <style>
-        html, body 
-        {
-            margin: 0;
-            height: 100%;
-        }
-        #c 
-        {
-            width: 100%;
-            height: 100%;
-            display: block;
-        }
-        div:focus {
-            outline:none;
-        }
-        .split {
-            position: absolute;
-            left: 0;
-            top: 0;
-            width: 100%;
-            height: 100%;
-            display: flex;
-        }
-        .split>div {
-            width: 100%;
-            height: 100%;
-        }
-    </style>
-</head>
-<body>
-    <div id="camera-position">Camera Position:</div>
-    <canvas id="c"></canvas>
-    <div class="split">
-        <div id="view1" tabindex="1">
-            <div id="blocker">
-                <div id="instructions">
-                    <p style="font-size:36px">
-                        Click to play
-                    </p>
-                    <p>
-                        Move: WASD or Arrow Keys<br/>
-                        Look: MOUSE
-                    </p>
-                </div>
-            </div>
-        </div>
-        <div id="view2" tabindex="2"></div>
-    </div>
-    <script type="importmap">
-        {
-            "imports": {
-                "three": "https://cdn.jsdelivr.net/npm/three@0.160.0/build/three.module.js",
-                "three/addons/": "https://cdn.jsdelivr.net/npm/three@0.160.0/examples/jsm/"
-            }
-        }
-    </script>
-    <script type="module" src="../public/js/cameraTester.js"></script>
-</body>
-</html>
