+++ pages/mainModel.html
@@ -7,9 +7,6 @@
 <!-- Carousel styles -->
 <link rel="stylesheet" href="../public/css/carousel.css">
 
-<!-- TIFF support library (optional - only if you have TIFF images) -->
-<script src="https://cdn.jsdelivr.net/npm/utif@3.1.0/UTIF.js"></script>
-
     <style>
         html,
         body {
@@ -461,7 +458,7 @@
             }
         }
     </script>
-    <script src="../public/js/pointerLockTest.js" type="module">
+    <script src="../public/js/main.js" type="module"></script>
     </script>
 </body>
