--- pages/uiBoxesTest.html
@@ -1,155 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>UI Boxes Test v2 - Virtual Fairgrounds</title>
-    <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
-    <style>
-        html,
-        body {
-            margin: 0;
-            height: 100%;
-            font-family: Arial, sans-serif;
-            background: linear-gradient(135deg, #417cbd 0%, #2a3579 100%);
-        }
-
-        #c {
-            width: 100%;
-            height: 100%;
-            display: block;
-            background: #222;
-        }
-
-        .test-controls {
-            position: fixed;
-            bottom: 20px;
-            left: 20px;
-            z-index: 200;
-        }
-
-        .test-button {
-            background: #417cbd;
-            color: white;
-            padding: 10px 20px;
-            border: none;
-            border-radius: 5px;
-            cursor: pointer;
-            margin-right: 10px;
-        }
-    </style>
-</head>
-
-<body>
-    <div id="c"></div>
-
-    <!-- Direct inline positioning to avoid any CSS conflicts -->
-    <div id="leftBox" style="
-        position: fixed;
-        left: 2vw;
-        top: 50vh;
-        transform: translateY(-50%);
-        width: 23vw;
-        height: 38vh;
-        background: rgba(42, 53, 121, 0.85);
-        border: 3px solid red;
-        border-radius: 10px;
-        padding: 20px;
-        color: white;
-        overflow-y: auto;
-        z-index: 100;
-    ">
-        <h3>Left Box (RED)</h3>
-        <p><strong>Position:</strong> Left side, vertically centered</p>
-        <p>Fixed position using top: 50vh + translateY(-50%)</p>
-    </div>
-
-    <div id="middleBox" style="
-        position: fixed;
-        left: 50vw;
-        transform: translateX(-50%);
-        bottom: 5vh;
-        width: 48vw;
-        height: 18vh;
-        background: rgba(42, 53, 121, 0.85);
-        border: 3px solid green;
-        border-radius: 10px;
-        padding: 20px;
-        color: white;
-        overflow-y: auto;
-        z-index: 100;
-    ">
-        <h3>Middle Box (GREEN) - BOTTOM POSITIONED</h3>
-        <p><strong>Position:</strong> Bottom center</p>
-        <p>Fixed position using bottom: 5vh + left: 50vw + translateX(-50%)</p>
-    </div>
-
-    <div id="rightBox" style="
-        position: fixed;
-        right: 2vw;
-        top: 50vh;
-        transform: translateY(-50%);
-        width: 23vw;
-        height: 38vh;
-        background: rgba(42, 53, 121, 0.85);
-        border: 3px solid blue;
-        border-radius: 10px;
-        padding: 20px;
-        color: white;
-        overflow-y: auto;
-        z-index: 100;
-    ">
-        <h3>Right Box (BLUE)</h3>
-        <p><strong>Position:</strong> Right side, vertically centered</p>
-        <p>Fixed position using right: 2vw + top: 50vh + translateY(-50%)</p>
-    </div>
-
-    <div class="test-controls">
-        <button class="test-button" onclick="logPositions()">Log Positions</button>
-        <button class="test-button" onclick="moveMiddleToTop()">Move Middle to Top</button>
-        <button class="test-button" onclick="moveMiddleToBottom()">Move Middle to Bottom</button>
-    </div>
-
-    <script>
-        function logPositions() {
-            const boxes = ['leftBox', 'middleBox', 'rightBox'];
-            boxes.forEach(id => {
-                const box = document.getElementById(id);
-                const rect = box.getBoundingClientRect();
-                console.log(`${id}:`, {
-                    left: rect.left,
-                    top: rect.top,
-                    right: rect.right,
-                    bottom: rect.bottom,
-                    width: rect.width,
-                    height: rect.height
-                });
-            });
-            console.log('Window height:', window.innerHeight);
-            console.log('Middle box should be near bottom. Bottom position:', document.getElementById('middleBox').getBoundingClientRect().bottom);
-        }
-
-        function moveMiddleToTop() {
-            const middle = document.getElementById('middleBox');
-            middle.style.bottom = 'auto';
-            middle.style.top = '5vh';
-            console.log('Moved middle to top');
-        }
-
-        function moveMiddleToBottom() {
-            const middle = document.getElementById('middleBox');
-            middle.style.top = 'auto';
-            middle.style.bottom = '5vh';
-            console.log('Moved middle to bottom');
-        }
-
-        // Auto-log on load
-        window.addEventListener('load', function () {
-            setTimeout(() => {
-                console.log('=== INITIAL POSITIONS ===');
-                logPositions();
-            }, 500);
-        });
-    </script>
-</body>
-
-</html>
