+++ public/js/config/locations.js
@@ -1,0 +1,33 @@
+// Interactive location definitions
+export const INTERACTION_ZONES = {
+    theater: {
+        position: { x: -32, y: 31, z: 7 },
+        radius: 8,
+        locationId: 'theater'
+    },
+    cleaners: {
+        position: { x: -35, y: 31, z: 32 },
+        radius: 4,
+        locationId: 'cleaners'
+    },
+    dominos: {
+        position: { x: -35.2, y: 31, z: 57.8 },
+        radius: 3,
+        locationId: 'dominos'
+    },
+    records: {
+        position: { x: -36, y: 31, z: 63 },
+        radius: 2,
+        locationId: 'records'
+    },
+    northEnd: {
+        position: { x: -50, y: 31, z: -28 },
+        radius: 11,
+        locationId: 'northEnd'
+    },
+    southEnd: {
+        position: { x: -52, y: 31, z: 86 },
+        radius: 11,
+        locationId: 'southEnd'
+    }
+};
