+++ public/js/utils/GUIHelpers.js
@@ -1,0 +1,74 @@
+/*
+  GUI helper for managing min/max values with constraints
+ */
+export class MinMaxGUIHelper {
+    constructor(obj, minProp, maxProp, minDif) {
+        this.obj = obj;
+        this.minProp = minProp;
+        this.maxProp = maxProp;
+        this.minDif = minDif;
+    }
+    get min() {
+        return this.obj[this.minProp];
+    }
+    set min(v) {
+        this.obj[this.minProp] = v;
+        this.obj[this.maxProp] = Math.max(this.obj[this.maxProp], v + this.minDif);
+        this.obj.updateProjectionMatrix(); // Update projection matrix when near changes
+    }
+    get max() {
+        return this.obj[this.maxProp];
+    }
+    set max(v) {
+        this.obj[this.maxProp] = v;
+        this.min = this.min;  // This calls the min setter above, which includes updateProjectionMatrix
+    }
+}
+
+// This is the color gui for the directional light
+class ColorGUIHelper {
+    constructor(object, prop) {
+        this.object = object;
+        this.prop = prop;
+    }
+    get value() {
+        return `#${this.object[this.prop].getHexString()}`;
+    }
+    set value(hexString) {
+        this.object[this.prop].set(hexString);
+    }
+}
+
+// This is the gui for fog to control legit everything
+class FogGUIHelper {
+    constructor(fog, camera) {
+        this.fog = fog;
+        this.camera = camera;
+    }
+    get near() {
+        return this.fog.near;
+    }
+    set near(v) {
+        this.fog.near = v;
+        this.fog.far = Math.max(this.fog.far, v);
+    }
+    get far() {
+        return this.fog.far;
+    }
+    set far(v) {
+        this.fog.far = v;
+        this.fog.near = Math.min(this.fog.near, v);
+    }
+    get density() {
+        return this.fog.density;
+    }
+    set density(v) {
+        this.fog.density = v;
+    }
+    get color() {
+        return `#${this.fog.color.getHexString()}`;
+    }
+    set color(hexString) {
+        this.fog.color.set(hexString);
+    }
+}
